+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Rename `compare_version` to `get_max_version` for readability
 
+- Add user metadata recipe
+
 ## [0.6.4] - 2022-04-11
 - bug fix in `default_create_and_send_custom_email` for emailverification recipe where we were not sending the email if env var was not set.
 - Fix telemetry issues related to asyncio when using FastAPI. Related issue: https://github.com/supertokens/supertokens-core/issues/421